+++ tests/full/test_full.tf
@@ -332,7 +332,6 @@ resource "test_assertions" "rtctrlSetASPathASN" {
     got         = data.aci_rest_managed.rtctrlSetASPathASN.content.order
     want        = "5"
   }
-
 }
 
 data "aci_rest_managed" "rtctrlSetNhUnchanged" {
@@ -348,7 +347,6 @@ resource "test_assertions" "rtctrlSetNhUnchanged" {
     got         = data.aci_rest_managed.rtctrlSetNhUnchanged.content.type
     want        = "nh-unchanged"
   }
-
 }
 
 data "aci_rest_managed" "rtctrlSetRedistMultipath" {
@@ -364,5 +362,4 @@ resource "test_assertions" "rtctrlSetRedistMultipath" {
     got         = data.aci_rest_managed.rtctrlSetRedistMultipath.content.type
     want        = "redist-multipath"
   }
-
 }